feat: update SCALE getters

+++ src/scale.h
@@ -88,7 +88,7 @@ typedef struct {
 
 
 uint8_t encode_scale(ScaleElem* encoded_value, uint8_t* value, size_t value_len, enum ScaleTypes type);
-uint8_t encode_composite_scale(ScaleElem* encoded_value, uint8_t *value, size_t value_len, const ScaleElem** elements, size_t elements_len, enum ScaleTypes type);
+uint8_t encode_composite_scale(ScaleElem* encoded_value, uint8_t *value, uint32_t value_len, const ScaleElem** elements, uint32_t elements_len, enum ScaleTypes type);
 void encode_scale_fixint_u8(ScaleElem *elem, uint8_t value);
 void encode_scale_fixint_u16(ScaleElem *elem, uint16_t value);
 void encode_scale_fixint_u32(ScaleElem *elem, uint32_t value);
@@ -97,14 +97,14 @@ uint8_t decode_scale(const ScaleElem *scale_elem, uint8_t *value, size_t *value_
 void print_scale(ScaleElem* scale_elem);
 uint8_t get_scale_value(const ScaleElem *scale_elem, uint8_t* value, uint8_t value_len);
 enum ScaleTypes get_scale_type(const ScaleElem *scale_elem);
-uint8_t get_scale_length(const ScaleElem *scale_elem);
+uint32_t get_scale_length(const ScaleElem *scale_elem);
 size_t get_vector_u8_size(size_t value_len);
 uint8_t as_scale_vector_u8(uint8_t* encoded_value, size_t encoded_value_len, uint8_t* value, size_t value_len);
 uint8_t contains_scale(const ScaleElem *el);
 uint8_t as_option(const ScaleElem *el, ScaleElem *el_option);
-size_t get_option_size(const ScaleElem *el);
-size_t get_vector_size(const ScaleElem** elements, uint32_t elements_len);
-size_t get_enumeration_size(const ScaleElem** elements, uint8_t elements_len);
+uint32_t get_option_size(const ScaleElem *el);
+uint32_t get_vector_size(const ScaleElem** elements, uint32_t elements_len);
+uint32_t get_enumeration_size(const ScaleElem** elements, uint8_t elements_len);
 void uint_as_compact(Compact *compact, uint32_t value);
 
 #define is_bigendian (!*(uint8_t *)&(uint16_t){1})
